Create data generator
Made a script to create a csv of cube scrambles with their solutions in order to train the model. Didn't have any particular issues coding it but it doesn't announce when it gets to 100% which is a bit annoying but not worth bothering to fix as it doesn't affect the functionality and it only needs to be used once.